Enum unic_ucd_category::GeneralCategory

pub enum GeneralCategory {
    UppercaseLetter,
    LowercaseLetter,
    TitlecaseLetter,
    ModifierLetter,
    OtherLetter,
    NonspacingMark,
    SpacingMark,
    EnclosingMark,
    DecimalNumber,
    LetterNumber,
    OtherNumber,
    ConnectorPunctuation,
    DashPunctuation,
    OpenPunctuation,
    ClosePunctuation,
    InitialPunctuation,
    FinalPunctuation,
    OtherPunctuation,
    MathSymbol,
    CurrencySymbol,
    ModifierSymbol,
    OtherSymbol,
    SpaceSeparator,
    LineSeparator,
    ParagraphSeparator,
    Control,
    Format,
    Surrogate,
    PrivateUse,
    Unassigned,
}

Represents the Unicode Character General_Category property.

This is a useful breakdown into various character types which can be used as a default categorization in implementations. For the property values, see General_Category Values.

Variants

UppercaseLetter

An uppercase letter

LowercaseLetter

A lowercase letter

TitlecaseLetter

A digraphic character, with first part uppercase

ModifierLetter

A modifier letter

OtherLetter

Other letters, including syllables and ideographs

NonspacingMark

A nonspacing combining mark (zero advance width)

SpacingMark

A spacing combining mark (positive advance width)

EnclosingMark

An enclosing combining mark

DecimalNumber

A decimal digit

LetterNumber

A letterlike numeric character

OtherNumber

A numeric character of other type

ConnectorPunctuation

A connecting punctuation mark, like a tie

DashPunctuation

A dash or hyphen punctuation mark

OpenPunctuation

An opening punctuation mark (of a pair)

ClosePunctuation

A closing punctuation mark (of a pair)

InitialPunctuation

An initial quotation mark

FinalPunctuation

A final quotation mark

OtherPunctuation

A punctuation mark of other type

MathSymbol

A symbol of mathematical use

CurrencySymbol

A currency sign

ModifierSymbol

A non-letterlike modifier symbol

OtherSymbol

A symbol of other type

SpaceSeparator

A space character (of various non-zero widths)

LineSeparator

U+2028 LINE SEPARATOR only

ParagraphSeparator

U+2029 PARAGRAPH SEPARATOR only

Control

A C0 or C1 control code

Format

A format control character

Surrogate

A surrogate code point

PrivateUse

A private-use character

Unassigned

Unassigned

Implementations

impl GeneralCategory

pub fn of(ch: char) -> GeneralCategory

Find the GeneralCategory of a single char.

impl GeneralCategory

pub fn is_cased_letter(&self) -> bool

Lu | Ll | Lt (Short form: LC)

pub fn is_letter(&self) -> bool

Lu | Ll | Lt | Lm | Lo (Short form: L)

pub fn is_mark(&self) -> bool

Mn | Mc | Me (Short form: M)

pub fn is_number(&self) -> bool

Nd | Nl | No (Short form: N)

pub fn is_punctuation(&self) -> bool

Pc | Pd | Ps | Pe | Pi | Pf | Po (Short form: P)

pub fn is_symbol(&self) -> bool

Sm | Sc | Sk | So (Short form: S)

pub fn is_separator(&self) -> bool

Zs | Zl | Zp (Short form: Z)

pub fn is_other(&self) -> bool

Cc | Cf | Cs | Co | Cn (Short form: C)